refactor(phase-5c): extract trackedFetch + createSessionSaver to src/boot/
desktop.ts is now 3,034 LOC (-661 from 3,695); 8 boot modules now
total 905 LOC of focused setup code under src/boot/.

- src/boot/tracked-fetch.ts (75 LOC) — runtime implementation of
  wp.desktop.fetch. Wraps native fetch() with per-window activity
  attribution; honors opts.windowId / opts.window / focused
  fallback. Same one legitimate raw-fetch escape hatch the lint
  rule already calls out.
- src/boot/session-saver.ts (133 LOC) — createSessionSaver()
  with the SESSION_SAVE_DEBOUNCE_MS constant. Owns the debounced
  + immediate save pipeline including the sendBeacon nonce-on-URL
  workaround for unload-time saves.
- desktop.ts: drops both inlined definitions; imports from the
  new modules.

All green: lint, tsc --noEmit, vitest (106 files / 936 tests).
